requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.25
coreapi==2.3.0
coreschema==0.0.4
dict.sorted==1.0.0
Django==1.11.1
django-celery==3.2.1
django-compat==1.0.14
django-embed-video==1.1.2
django-filter==1.1.0
django-material==1.1.1
django-user-agents==0.3.2
djangorestframework==3.8.2
djangorestframework-bulk==0.2.1
drf-toolbox==0.1.7
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.7.1
itypes==1.1.0
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
num2words==0.5.4
openapi-codec==1.3.1
packaging==16.8
pyfcm==1.4.5
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
requests==2.14.2
requests-toolbelt==0.8.0
simplejson==3.10.0
six==1.10.0
twitter==1.17.1
tzlocal==1.4
ua-parser==0.7.3
uritemplate==3.0.0
user-agents==1.1.0
vine==1.1.3
django-jet==1.0.6
google-api-python-client==1.4.1
git+https://github.com/ItsCalebJones/django-rest-framework-docs.git